ppc64le: Support for ppc64le with nvidia-docker 2.0
Currently elftoolchain 0.7.1 does not support ppc64le elf headers.
The immediate workaround for this issue is to use the libelf-dev on
debian, or elfultils-libelf-devel on RedHat.  When using the native
libelf libraries, we avoid having to build elftoolchain.

However, Elf_Note is not defined anywhere except in the header
elfdefinitions.h, which currently resides in the elftoolchain
project space.  The struct has been moved into elftool.h located
locally in the src directory.