Compile libelf for aarch64
This is a bit of a workaround as we have to override the script libelf
uses to detect which platform it runs on.
libELF however specifies that for the version we are on (0.7.1) it
understands aarch64 binaries and can be compiled for aarch64.
See: https://sourceforge.net/p/elftoolchain/wiki/Releases-Release0.7.1/#release-notes-for-elftoolchain-software-version-071