Add ability to check for MIG mode on a given device
Note that this implementation should work for older versions of CUDA
(that don't have the MIG API calls), as well as for GPU brands which
don't support MIG. In both cases, we return successfully with an
indication that MIG is simply disabled. Only if a card is MIG capable
and returns an actual error from the call to get the MIG mode, do we
also return an error.

Signed-off-by: Kevin Klues <kklues@nvidia.com>