Rework CLI to support MIG device strings as input
For now the only command that uses this information is the 'list'
command. The configure command honors device strings for MIG devices,
but only currently continues to inject the device node for the MIG
device, and does nothing about the /proc access files for the GPU and
Compute instances. This support will be added in a later commit.

Signed-off-by: Kevin Klues <kklues@nvidia.com>