Use an environment variable to set the git revision
When packaging, you cannot expect to have the .git since the released
tarball does not contain that. Thus, it is needed to be able to use an
environment variable to set the git revision.

Note, if you do not set that variable, you still get the revision
running the git command.

Signed-off-by: Jordi Massaguer Pla <jmassaguerpla@suse.de>