Rework build system to cross-build for amd64, x86_64, and ppc64le
The cross builds rely on binfmt-support and qemu-user-static being
installed on your build machine in order to work.

On Ubuntu, for example, this is a simple matter of installing the
`binfmt-support` and `qemu-user´-static` packages and running the
following command to ensure the proper flags are set for your qemu
environment when performing the corss builds.

docker run --rm --privileged multiarch/qemu-user-static --reset -p yes -c yes

Signed-off-by: Kevin Klues <kklues@nvidia.com>