Remove 'ngx-updater' from being injected as part of the 'ngx' capability
It was decided that this is not something that should actually be
injected into a container (i.e. it's only appropriate to have at the
host level since it is used to update libraries and such on the fly). We
can revisit this again in the future if it is later decided that we
actually want to include it.