Update debian based dockerfiles to set distribution in changelog
The 'prepare' script looks for an envvar called DISTRIB and overwrites
any instances of the workd UNRELEASED in the changelog with this value.

According to https://www.debian.org/doc/manuals/developers-reference/pkgs.html#picking-a-distribution
this should be overwritten with the name of the debain distribution
(which is easily obtained from calling `lsb_release -c -s`).

This commit patches things up to ensure that we set the DISTRIB envvar
correctly and update the changelog accordingly before distribution.

Signed-off-by: Kevin Klues <kklues@nvidia.com>