Add setup_mig_minor_cgroups() to enable cgroups for MIG devices on a GPU
This is necessary to make the mig-confg and mig-monitor capabilities
useful. With /proc based capabilities all we had to do was bind-mount a
directory from /proc in order to give the container access to _all_
capabilities from that subtree down. This is required to ensure that a
container has access to any GPU devices that are reated / destroyed
after the container has already started.

With /dev based capabilities it is not as straight-forward. We now have
to enumerate all possible capabilities under a subtree and give the
container acess to each of them individually. We have designed it so
that the minor ids for the set of capabilities that exist under any
given tree in /proc is predetermined and stored in a file under
/proc/nvidia-caps/*-minors. We use the file mig-minors for our purposes
here.

Signed-off-by: Kevin Klues <kklues@nvidia.com>