Remove path_join() with already chrooted directory
In the recent patch to run mknod() over all of the nvidia capability
device nodes, we erroneously joined the relative path of the /proc files
with the root path of the driver installation.

Code higher up already chroots into the driver installation path,
making the join unnecessary (and actually incorrect).

This patch removes that join and just uses the constructed path directly.

Signed-off-by: Kevin Klues <kklues@nvidia.com>