Invoke docker make file separately
The files Makefile and mk/common.mk contain a number of checks that
are not portable. These should ideally be run INSIDE the docker images
defined in `mk/Dockerfile.*`. This change removes the include of
mk/docker.mk into the top-level makefile and allows the docker build
commands to be invoked using:

make -f mk/docker.mk ubuntu18.04

(as an example)

This is required to be able to build in Docker in Docker in Blossom.

Signed-off-by: Evan Lezar <elezar@nvidia.com>