Added libnvidia-nscq.so as an optional injected utility lib
The NSCQ library provides topology information of the NVSwitches and
GPUs to clients of the library such as DCGM. Note that currently, DCGM
is the only client of NSCQ. In the future, NSCQ will include a public
API for gathering NVSwitch information.

At present this is not part of the actual driver disstribution, but it
is revlocked to the driver with a suffix of e.g. "450.80.02" like other
driver libraries. However, in the near future the library package will
be installed as part of the driver. We are just getting ahead of that.

Signed-off-by: Kevin Klues <kklues@nvidia.com>