Mount all of /proc/.../nvidia/capabilities with --mig-{config, monitor}
After a long conversation with the driver team, it was concluded that
the current method of selectively injecting just the
/proc/driver/nvidia/capabilities/gpu* folders is not going to work.
These folders are recreated across a MIG mode change and therefore any
containers that have them bind mounted in will lose their reference to
them when this happens.

That said, this original design was based on the fact that
nvidia-capabilities were /proc based and we had to selective inject them
this way to prevent unauthorized access to certain capabilities.

Now that nvidia-capabilities are /dev based, it is no longer an issue to
allow a container to see other GPUs under
/proc/driver/nvidia/capabilities/gpu* that it may not actually have
access to.

As such, we are relaxing the constraint of only selectively injecting
/proc/driver/nvidia/capabilities/gpu* for the GPUs matching the list of
GPUs in --devices, and instead injecting the entire
/proc/driver/nvidia/capabilities folder so that as any folders/files
underneath that get created / deleted on the host – the container will
see those changes.

Note this is only for containers with --mig-config or --mig-monitor
privileges, and not just normal containers. Normal containers will
continue to only have the access files for the MIG devices they actually
have access to injected at runtime.

Signed-off-by: Kevin Klues <kklues@nvidia.com>