Allow for --platform arguments to be added to docker commands
This changes adds support for a make variable ADD_DOCKER_PLATFORM_ARGS that when
set adds the --platform=linux/${ARCH} command line arguments to the docker build
and docker run commands. This allows for cross platform builds on versions of
docker that support docker buildx.

This allows for the control required to build on macos and blossom while ensuring
compatibility (by default) with existing build systems such as gitlab CI.

Signed-off-by: Evan Lezar <elezar@nvidia.com>