Add nvcgo library that can be called from C for running Go code
Building of this library is gated with a WITH_NVCGO build variable.

At present, this library only contains logic to manipulate the devices cgroup.
It can handle both cgroupv1 and cgroupv2. This library is not yet linked or
loaded and no functions from this library are called / utilized anywhere in the
code yet.

Signed-off-by: Kevin Klues <kklues@nvidia.com>