Add new cgroup.c file to support both cgroupv1 and cgroupv2
The code in this file calls out to Go code in nvcgo library via RPC. The code
itself will only be compiled / linked in if the WITH_NVCGO make variable is set.
When this make variable is not set, the old legacy cgroup code from
cgroup_legacy.c will be included instead.

At present, this will error out with a permission error when attempting to
modify the device cgroups. A subsequent commit will patch things up to
explicitly set the right permissions across the call to set the device rules.

Signed-off-by: Kevin Klues <kklues@nvidia.com>