Move logic to chroot and drop privileges from RPC init into driver init
Only the driver RPC code requires a chroot in order to access the dri er
libraries so this doesn't make sense as part of the standard RPC library.
Likewise, dropping privileges should not be tied to the general RPC mechanism,
but rather something chosen by the individual service itself (in this case the
driver).

Signed-off-by: Kevin Klues <kklues@nvidia.com>