Add -fplan9-extensions to Makefile for future use
The only extension we plan to use is the ability to "extend" structs by
embeddeing other structs in them.

For example:
struct foo {
        int a;
};
struct bar {
        struct foo;
        int b;
};

Is now legal with these extensions, and an instance of bar can access the 'a'
field as if it was directly part of the bar struct.

Signed-off-by: Kevin Klues <kklues@nvidia.com>