Fix permissions when mounting GSP firmware
This change addresses inconsistencies in the mounting of the GSP firmware
for newer devices. This removes the mount of the entire firmware directory
and insteade mounts the gsp.bin file directly. This allows the file to
be executable (as it is on the host) and also ensures that it is mounted
read-only instead of read-write.

Signed-off-by: Evan Lezar <elezar@nvidia.com>