Fix bug when running host ldconfig using glibc compiled with a non-standard prefix
This change explicitly sets the ldcache and config paths when running ldconfig
to the defaults of "/etc/ld.so.cache" and "/etc/ld.so.conf". This handles the fact that
the host ldconfig may be expecting another path due to glibc being
compiled with a prefix.

See https://gitlab.com/nvidia/container-toolkit/libnvidia-container/-/issues/1

Signed-off-by: Evan Lezar <elezar@nvidia.com>