Remove libnvidia-container0 dependency
This change removes the dependency on libnvidia-container0
on ubuntu-based arm64 systems. The library is required to support
the legacy mode on Tegra-based systems. This has not been the
default since the 1.10.0 release of the NVIDIA Container Toolkit.

If the legacy mode is required, the libnvidia-container0 packages
can be installed separately from the libnvidia-container-tools
pacakge.

Signed-off-by: Evan Lezar <elezar@nvidia.com>