Require version match between libnvidia-container-tools and libnvidia-container1
This change ensures that the package versions for the libnvidia-container-tools
and libnvidia-contianer1 packages match. On rpm-based systems, this will
mean that if the version of libnvidia-container-tools is specified the same
version is used for the libnvidia-container1 package.

On deb-based systems, an error is shown if the version of libnvidia-container1
is not specified and a newer version is available from the package repository.

Signed-off-by: Evan Lezar <elezar@nvidia.com>