Fix pointer accessing local variable out of scope
The C11 standard states that the value of a pointer is undefined
outside of its lifetime. Since we were initilizing the argv pointer
in two conditional blocks, compilers that implement this standard were
cleaning up the value and this caused invalid dereferences of the
pointer.

The change introduces two local variables with the same lifetime of the
argv pointer and assigns these in the conditional blocks instead.

Signed-off-by: Evan Lezar <elezar@nvidia.com>